fix: unify setup snapshot validation
